+++ R/covsel.R
@@ -1,10 +1,11 @@
-covsel <- function(X, Y, nvar = NULL, scaly = TRUE, weights = NULL) {
+covsel <- function(X, Y, nvar = NULL, Xscaling = c("none", "pareto", "sd")[1], Yscaling = c("none", "pareto", "sd")[1], weights = NULL) {
   
     X <- .mat(X)
+    varnames <- colnames(X)
     zdim <- dim(X)
     n <- zdim[1]
     p <- zdim[2]
-    Y <- .mat(Y, "x")
+    Y <- .mat(Y, "y")
     q <- dim(Y)[2]
   
     if(is.null(nvar)) 
@@ -15,12 +16,17 @@ covsel <- function(X, Y, nvar = NULL, scaly = TRUE, weights = NULL) {
     weights <- .mweights(weights)
   
     xmeans <- .colmeans(X, weights = weights)
-    X <- .center(X, xmeans)
     ymeans <- .colmeans(Y, weights = weights)
-    Y <- .center(Y, ymeans)
-  
-    if(scaly)
-        Y <- .scale(Y, rep(0, q), sqrt(colSums(weights * Y * Y)))
+    
+    if(Xscaling == "none") {xscales <- rep(1, p)}
+    if(Yscaling == "none") {yscales <- rep(1, q)}
+    if(Xscaling == "sd") {xscales <- sqrt(.colvars(X, weights = weights))}
+    if(Yscaling == "sd") {yscales <- sqrt(.colvars(Y, weights = weights))}
+    if(Xscaling == "pareto") {xscales <- sqrt(sqrt(.colvars(X, weights = weights)))}
+    if(Yscaling == "pareto") {yscales <- sqrt(sqrt(.colvars(Y, weights = weights)))}
+    
+    X <- scale(X, center = xmeans, scale = xscales)
+    Y <- scale(Y, center = ymeans, scale = yscales)
   
     xsstot <- sum(weights * X * X)
     ysstot <- sum(weights * Y * Y)
@@ -44,9 +50,12 @@ covsel <- function(X, Y, nvar = NULL, scaly = TRUE, weights = NULL) {
     cumpvarx <- 1 - xss / xsstot
     cumpvary <- 1 - yss / ysstot
     sel <- data.frame(sel = selvar, 
-                      cumpvarx = cumpvarx, cumpvary = cumpvary)
+                      cumpvarx = cumpvarx, cumpvary = cumpvary,
+                      selnames = sapply(selvar, function(j) varnames[j]))
   
-    list(sel = sel, weights = weights)
-
+    structure(
+      list(sel = sel, weights = weights),# xmeans = xmeans, ymeans = ymeans, xscales = xscales, yscales = yscales)
+      class = c("Covsel"))
     }
+