+++ inst/NEWS.md
@@ -15,3 +15,8 @@
 - New functions : plsr_plsda_allsteps; mbplsr_mbplsda_allsteps; soplsr_soplsda_allsteps
 - Function changes : changes in the cross-validation partition to avoid storing sub-matrices in soplsrcv, soplsrdacv, soplsldacv, soplsqdacv functions
 - bugfix R/gridcv.R; R/gridcvlv.R; R/gridcvlb.R; R/gridscore.R; R/gridscorelv.R; R/gridscorelb.R.
+
+## **Version 0.1-4**
+
+- New functions : covsellm, covselrda, covsellda, covselqda
+- Function changes : "Xscaling" and "Yscaling" arguments added to covsel instead of "scaly"